backend/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
motor
dexscreener
zksync2
python-decouple